test(computer): isolate L4 device-abort test into its own file
The previous commit put the L4 regression as a second describe block in
playground-cancel.test.ts, but the first describe's afterAll calls
agent.destroy() which tears down the Xvfb instance on display :99. The
second describe's device.connect() then crashed the vitest worker with
XIO: fatal IO error 2 on X server ":99", and vitest treated the worker
crash as a 0 exit code so CI silently went green — defeating the whole
point of the demo-red commit.

Split L4 into its own test file so it runs in its own vitest worker
with its own Xvfb. Update the headless-linux-cancel workflow to run
both files.

L4 fix is still intentionally disabled in device.ts (carried over from
the previous commit) so this run should produce a real red. The next
commit restores L4.
